Added a verbose flag + started splitting purposes into files.
